requirements.txt
annotated-types==0.7.0
anyio==4.9.0
azure-communication-callautomation==1.4.0
azure-communication-identity==1.5.0
azure-communication-sms==1.1.0
azure-core==1.34.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
fastapi==0.115.13
h11==0.16.0
idna==3.10
isodate==0.7.2
msrest==0.7.1
oauthlib==3.3.0
pydantic==2.11.7
pydantic-core==2.33.2
python-dotenv==1.1.0
requests==2.32.4
requests-oauthlib==2.0.0
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
typing-extensions==4.14.0
typing-inspection==0.4.1
urllib3==2.4.0
uvicorn==0.34.3
